refactor: Move Redis workers code to main controller
- Move _get_site_config_key_value from workers.py to rq_controller.py
- Add new helper functions _get_redis_connection and _get_worker_states
- Clean up imports and remove redundant code
- Delete workers.py module
- Simplify error handling and connection management
